% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\docType{data}
\name{ab}
\alias{ab}
\title{Alberta Data}
\format{
An object of class \code{data.frame}
}
\usage{
data(ab)
}
\description{
Data from the 2018 Calgary: The Road Ahead Survey conducted
by Janet Brown Opinion Research / Trend Research on behalf of CBC Calgary.
The data are comprised of n=1,200 adults of the Alberta general population
from a stratified (Calgary/Edmonton/Other AB) RDD sample frame. Interviews
were conducted between March 13 to April 5, 2018.
}
\keyword{datasets}
