% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTab.R
\name{cTab.data.frame}
\alias{cTab.data.frame}
\title{cTab.data.frame}
\usage{
\method{cTab}{data.frame}(design, yvar, xvar, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvar}{String denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
cTab.data.frame
}
