% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctab}
\alias{ctab}
\title{ctab}
\usage{
ctab(design, yvar, xvar, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvar}{String denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A tibble
}
\description{
ctab
}
\examples{
library(survey)
data("ab")
absvy <- svydesign(~1, data = ab, weights = ~weight)
ctab(absvy, yvar = "dvote3", xvar = "region")

}
\author{
John Santos
}
