% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTab.R
\name{cTabs.data.frame}
\alias{cTabs.data.frame}
\title{cTabs.data.frame}
\usage{
\method{cTabs}{data.frame}(design, yvar, xvars, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvars}{String (or vector of strings) denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
cTabs.data.frame
}
