% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTab.R
\name{cTabs.survey.design}
\alias{cTabs.survey.design}
\title{cTabs.survey.design}
\usage{
\method{cTabs}{survey.design}(design, yvar, xvars, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvars}{String (or vector of strings) denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
cTabs.survey.design
}
