% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctabs_format}
\alias{ctabs_format}
\title{format_ctabs}
\usage{
ctabs_format(obj, ...)
}
\arguments{
\item{obj}{An object of class "ctabs.test" created with the \code{testcols} function.}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A tibble

A tibble
}
\description{
format_ctabs
}
\author{
John Santos
}
