% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctabs_kbl.data.frame}
\alias{ctabs_kbl.data.frame}
\title{Print objects outputted by \code{format_ctabs} using \code{kable}}
\usage{
\method{ctabs_kbl}{data.frame}(obj, xvarnames = NULL, ...)
}
\arguments{
\item{obj}{An list created with \code{testcols} or a data frame created with
\code{format_ctabs}.}

\item{xvarnames}{A vector of character strings that replaces the names of your
column variables. (Default \code{NULL}; i.e. the function uses the variable
names as they appear in \code{R}.) Useful if you want to specify nicely
formatted variable names..}

\item{...}{Other arguments (not currently implemented).}
}
\value{
A formatted \code{kable} table
}
\description{
Print objects outputted by \code{format_ctabs} using \code{kable}
}
\author{
John Santos
}
