% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbsMisc_functions.R
\name{describe}
\alias{describe}
\title{Detailed summary statistics}
\usage{
describe(x, digits = 4, convertFactors = FALSE, ...)
}
\arguments{
\item{x}{Vector on which to calculate summary statistics}

\item{digits}{SCalar indicating how many decimal points to return}

\item{convertFactors}{Boolean indicating whether to convert factors to numeric}

\item{...}{Other arguments (not currently implemented)}
}
\description{
Detailed summary statistics
}
