% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbsMisc_functions.R
\name{ggfontsize}
\alias{ggfontsize}
\title{ggfontsize}
\usage{
ggfontsize(basesize = 11, angle.labs = FALSE)
}
\arguments{
\item{basesize}{Scalar denoting the base size (in points) of plot text}

\item{angle.labs}{Boolean indicating whether to angle x-axis labels}
}
\description{
ggfontsize
}
