% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbspal.R
\docType{data}
\name{jbspal}
\alias{jbspal}
\title{jbsMisc colour palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
data(jbspal)
}
\description{
jbsMisc colour palettes
}
\keyword{datasets}
