% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbsMisc_functions.R
\name{moeProp}
\alias{moeProp}
\title{Margin of error of proportion}
\usage{
moeProp(prop, n, ci = 0.95, ...)
}
\arguments{
\item{prop}{Proportion for which margin of error is to be calculated  (in decimal format)}

\item{n}{Sample size}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A scalar denoting the SE
}
\description{
Margin of error of proportion
}
