% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{print.ctabs}
\alias{print.ctabs}
\title{Print method for ctabs objects}
\usage{
\method{print}{ctabs}(x, digits = 0, ...)
}
\arguments{
\item{x}{An object of class "ctabs" produced by \code{ctabs()}}

\item{digits}{Scalar indicating how many decimal places to display (default 0)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
Prints the summary crosstab table from a ctabs list object. It
automatically converts decimals to percentages.
}
