% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{print.ctabs.test}
\alias{print.ctabs.test}
\title{Print method for ctabs.test objects}
\usage{
\method{print}{ctabs.test}(x, ...)
}
\arguments{
\item{x}{An object of class "ctabs.test" produced by \code{ctabs()}}

\item{...}{Other arguments (not currently implemented)}
}
\description{
Prints dataframes within ctabs.test list object and adds column
letters marking significantly different column proportions.
}
