% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{print.ctabs.test_}
\alias{print.ctabs.test_}
\title{Print method for ctabs.test_ objects}
\usage{
\method{print}{ctabs.test_}(x, ...)
}
\arguments{
\item{x}{An object of class "ctabs.test" produced by \code{ctabs()}}

\item{...}{Other arguments (not currently implemented)}
}
\description{
Prints only the summary tables from a ctabs.test_ list object.
}
