% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbsMisc_functions.R
\name{seProp}
\alias{seProp}
\title{Calculate SE of proportion}
\usage{
seProp(prop, n, ...)
}
\arguments{
\item{prop}{Proportion for which SE is to be calculated (in decimal format)}

\item{n}{Sample size}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A scalar denoting the SE
}
\description{
Calculate SE of proportion
}
