% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTab.R
\name{testCols.cTabs.list}
\alias{testCols.cTabs.list}
\title{testCols.cTabs.list}
\usage{
\method{testCols}{cTabs.list}(obj, adj.p = FALSE, ...)
}
\arguments{
\item{obj}{An object created with the cTab or Ctabs functions}

\item{adj.p}{Boolean indicating whether to adjust for multiple comparisons}

\item{...}{Other arguments (not currently implemented)}
}
\description{
testCols.cTabs.list
}
