% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{testcols.ctabs}
\alias{testcols.ctabs}
\title{testcols.ctabs}
\usage{
\method{testcols}{ctabs}(obj, adj.p = TRUE, ...)
}
\arguments{
\item{obj}{An object created with the ctab or Ctabs functions}

\item{adj.p}{Boolean indicating whether to adjust for multiple comparisons}

\item{...}{Other arguments (not currently implemented)}
}
\description{
testcols.ctabs
}
