% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTab.R
\name{testCols}
\alias{testCols}
\title{testCols}
\usage{
testCols(obj, adj.p = FALSE, ...)
}
\arguments{
\item{obj}{An object created with the cTab or Ctabs functions}

\item{adj.p}{Boolean indicating whether to adjust for multiple comparisons}

\item{...}{Other arguments (not currently implemented)}
}
\description{
testCols
}
