% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{testcols_.ctabs}
\alias{testcols_.ctabs}
\title{testcols_.ctabs}
\usage{
\method{testcols_}{ctabs}(obj, adj.p = TRUE, siglevel = 0.05, ...)
}
\arguments{
\item{obj}{An object of class "ctabs" created with the \code{ctabs()} function}

\item{adj.p}{Boolean indicating whether to adjust for multiple comparisons}

\item{siglevel}{Numeric indicating significance level to use (default .05)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
testcols_.ctabs
}
