% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{testcols_}
\alias{testcols_}
\title{testcols_}
\usage{
testcols_(obj, adj.p = FALSE, siglevel = 0.05, ...)
}
\arguments{
\item{obj}{An object of class "ctabs" created with the \code{ctabs()} function}

\item{adj.p}{Boolean indicating whether to apply Holm correction to p-values (default TRUE)}

\item{siglevel}{Numeric indicating significance level to use (default .05)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
testcols_
}
\examples{
library(survey)
data("ab")
absvy <- svydesign(~1, data = ab, weights = ~weight)
table_vote1 <- ctabs(absvy, yvar = "dvote3", xvars = c("region", "female", "age"))
table_vote1_tests <- testcols_(table_vote1)
table_vote1_tests

}
