% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbsMisc_functions.R
\name{testProps}
\alias{testProps}
\title{Title}
\usage{
testProps(a.prop, a.n, b.prop, b.n, ...)
}
\arguments{
\item{a.prop}{Proportion for Group A (in decimal format)}

\item{a.n}{Size of Group A}

\item{b.prop}{Proportion for Group B (in decimal format)}

\item{b.n}{Size of Group B}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A data frame
}
\description{
Title
}
