% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbsMisc_functions.R
\name{unWeight}
\alias{unWeight}
\title{Unweight survey design}
\usage{
unWeight(obj, ...)
}
\arguments{
\item{obj}{A survey design object}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A survey design object
}
\description{
Sets sampling weights to 1. This "unweights" a survey design object, but
keeps it as a survey design object. Written to help with develop the
svyEffects package.
}
