% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctab.data.frame}
\alias{ctab.data.frame}
\title{ctab.data.frame}
\usage{
\method{ctab}{data.frame}(design, yvar, xvar, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvar}{String denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
ctab.data.frame
}
