% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctabs.survey.design}
\alias{ctabs.survey.design}
\title{ctabs.survey.design}
\usage{
\method{ctabs}{survey.design}(design, yvar, xvars, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvars}{String (or vector of strings) denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A list of class \code{ctabs} with dataframes:
\describe{
\item{\code{$Summary}}{The summary crosstab of every column variable
against the row variable}
\item{\code{$Total}}{The marginal frequencies and proportions of the row variable}
\item{\code{$(name)}}{A grouped dataframe of the crosstab of the respective
column variable against the row variable}
}
}
\description{
ctabs.survey.design
}
\examples{
library(survey)
data("ab")
absvy <- svydesign(~1, data = ab, weights = ~weight)
table_vote1 <- ctabs(absvy, yvar = "dvote3", xvars = c("region", "female", "age"))
table_vote1

}
\author{
John Santos
}
