% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctabs.survey.design}
\alias{ctabs.survey.design}
\title{ctabs.survey.design}
\usage{
\method{ctabs}{survey.design}(design, yvar, xvars, ...)
}
\arguments{
\item{design}{Survey design object or data frame}

\item{yvar}{String denoting name of row variable (in quotes)}

\item{xvars}{String (or vector of strings) denoting name of column variable (in quotes)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
ctabs.survey.design
}
