% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{plot.ctabs}
\alias{plot.ctabs}
\title{plot method for ctabs objects}
\usage{
\method{plot}{ctabs}(x, dodge = FALSE, ...)
}
\arguments{
\item{x}{Object of class "ctabs" created by \code{ctabs()}}

\item{dodge}{Boolean indicating whether bars should be dodged (by default, they are stacked)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
plot method for ctabs objects
}
