% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{testcols.ctabs}
\alias{testcols.ctabs}
\title{testcols.ctabs}
\usage{
\method{testcols}{ctabs}(obj, adj.p = TRUE, siglevel = 0.05, ...)
}
\arguments{
\item{obj}{An object of class "ctabs" created with the \code{ctabs()} function}

\item{adj.p}{Boolean indicating whether to adjust for multiple comparisons}

\item{siglevel}{Numeric indicating significance level to use (default .05)}

\item{...}{Other arguments (not currently implemented)}
}
\value{
A list of class \code{ctabs.test} with one dataframe per crosstab of
the respective column variable against the row variable. Each crosstab has
frequencies, column proportions, and letters indicating significant differences
of proportions between pairs of columns.
}
\description{
This is (mainly) a wrapper around \code{rstatix::row_wise_prop_test}.
It takes a \code{ctabs} list of crosstabs, tests for all pairwise differences
and outputs a list of data frames that merge the frequencies/percentages with
significantly different column proportions being marked with letters.
}
\author{
John Santos
}
