% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{testcols}
\alias{testcols}
\title{testcols}
\usage{
testcols(obj, adj.p = FALSE, ...)
}
\arguments{
\item{obj}{An object of class "ctabs" created with the \code{ctabs()} function}

\item{adj.p}{Boolean indicating whether to apply Holm correction to p-values (default TRUE)}

\item{...}{Other arguments (not currently implemented)}
}
\description{
testcols
}
